+++ src/bluer/device.rs
@@ -1,8 +1,11 @@
-use futures_util::StreamExt;
-use tokio::pin;
+use std::sync::Arc;
 
-use super::adapter::session;
+use futures_core::Stream;
+use futures_lite::StreamExt;
+
+use super::l2cap_channel::{L2capChannelReader, L2capChannelWriter};
 use super::DeviceId;
+use crate::device::ServicesChanged;
 use crate::error::ErrorKind;
 use crate::pairing::PairingAgent;
 use crate::{btuuid, AdvertisementData, Device, Error, ManufacturerData, Result, Service, Uuid};
@@ -10,7 +13,8 @@ use crate::{btuuid, AdvertisementData, Device, Error, ManufacturerData, Result, 
 /// A Bluetooth LE device
 #[derive(Debug, Clone)]
 pub struct DeviceImpl {
-    pub(super) inner: bluer::Device,
+    pub(super) inner: Arc<bluer::Device>,
+    session: Arc<bluer::Session>,
 }
 
 impl PartialEq for DeviceImpl {
@@ -35,9 +39,10 @@ impl std::fmt::Display for DeviceImpl {
 }
 
 impl Device {
-    pub(super) fn new(adapter: &bluer::Adapter, addr: bluer::Address) -> Result<Device> {
+    pub(super) fn new(session: Arc<bluer::Session>, adapter: &bluer::Adapter, addr: bluer::Address) -> Result<Device> {
         Ok(Device(DeviceImpl {
-            inner: adapter.device(addr)?,
+            inner: Arc::new(adapter.device(addr)?),
+            session,
         }))
     }
 }
@@ -106,53 +111,71 @@ impl DeviceImpl {
             // `agent` to the static lifetime.
             let agent: &'static T = unsafe { std::mem::transmute(agent) };
 
-            async fn req_device(adapter: &str, addr: bluer::Address) -> Result<Device, bluer::agent::ReqError> {
-                let session = session().await.map_err(|_| bluer::agent::ReqError::Rejected)?;
+            async fn req_device(
+                session: Arc<bluer::Session>,
+                adapter: &str,
+                addr: bluer::Address,
+            ) -> Result<Device, bluer::agent::ReqError> {
                 let adapter = session.adapter(adapter).map_err(|_| bluer::agent::ReqError::Rejected)?;
                 let device = adapter.device(addr).map_err(|_| bluer::agent::ReqError::Rejected)?;
-                Ok(Device(DeviceImpl { inner: device }))
+                Ok(Device(DeviceImpl {
+                    inner: Arc::new(device),
+                    session,
+                }))
             }
 
             bluer::agent::Agent {
-                request_passkey: Some(Box::new(move |req: bluer::agent::RequestPasskey| {
-                    Box::pin(async move {
-                        let device = req_device(&req.adapter, req.device).await?;
-                        match agent.request_passkey(&device).await {
-                            Ok(passkey) => Ok(passkey.into()),
-                            Err(_) => Err(bluer::agent::ReqError::Rejected),
-                        }
-                    })
+                request_passkey: Some(Box::new({
+                    let session = self.session.clone();
+                    move |req: bluer::agent::RequestPasskey| {
+                        let session = session.clone();
+                        Box::pin(async move {
+                            let device = req_device(session, &req.adapter, req.device).await?;
+                            match agent.request_passkey(&device).await {
+                                Ok(passkey) => Ok(passkey.into()),
+                                Err(_) => Err(bluer::agent::ReqError::Rejected),
+                            }
+                        })
+                    }
                 })),
-                display_passkey: Some(Box::new(move |req: bluer::agent::DisplayPasskey| {
-                    Box::pin(async move {
-                        let device = req_device(&req.adapter, req.device).await?;
-                        if let Ok(passkey) = req.passkey.try_into() {
-                            agent.display_passkey(&device, passkey);
-                            Ok(())
-                        } else {
-                            Err(bluer::agent::ReqError::Rejected)
-                        }
-                    })
+                display_passkey: Some(Box::new({
+                    let session = self.session.clone();
+                    move |req: bluer::agent::DisplayPasskey| {
+                        let session = session.clone();
+                        Box::pin(async move {
+                            let device = req_device(session, &req.adapter, req.device).await?;
+                            if let Ok(passkey) = req.passkey.try_into() {
+                                agent.display_passkey(&device, passkey);
+                                Ok(())
+                            } else {
+                                Err(bluer::agent::ReqError::Rejected)
+                            }
+                        })
+                    }
                 })),
-                request_confirmation: Some(Box::new(move |req: bluer::agent::RequestConfirmation| {
-                    Box::pin(async move {
-                        let device = req_device(&req.adapter, req.device).await?;
-                        if let Ok(passkey) = req.passkey.try_into() {
-                            agent
-                                .confirm_passkey(&device, passkey)
-                                .await
-                                .map_err(|_| bluer::agent::ReqError::Rejected)
-                        } else {
-                            Err(bluer::agent::ReqError::Rejected)
-                        }
-                    })
+                request_confirmation: Some(Box::new({
+                    let session = self.session.clone();
+                    move |req: bluer::agent::RequestConfirmation| {
+                        let session = session.clone();
+                        Box::pin(async move {
+                            let session = session.clone();
+                            let device = req_device(session, &req.adapter, req.device).await?;
+                            if let Ok(passkey) = req.passkey.try_into() {
+                                agent
+                                    .confirm_passkey(&device, passkey)
+                                    .await
+                                    .map_err(|_| bluer::agent::ReqError::Rejected)
+                            } else {
+                                Err(bluer::agent::ReqError::Rejected)
+                            }
+                        })
+                    }
                 })),
                 ..Default::default()
             }
         };
 
-        let session = session().await?;
-        let _handle = session.register_agent(agent).await?;
+        let _handle = self.session.register_agent(agent).await?;
 
         self.pair().await
     }
@@ -163,8 +186,7 @@ impl DeviceImpl {
             self.inner.disconnect().await?;
         }
 
-        let session = session().await?;
-        let adapter = session.adapter(self.inner.adapter_name())?;
+        let adapter = self.session.adapter(self.inner.adapter_name())?;
         adapter.remove_device(self.inner.address()).await.map_err(Into::into)
     }
 
@@ -174,41 +196,56 @@ impl DeviceImpl {
     }
 
     /// Discover the primary service(s) of this device with the given [`Uuid`].
-    pub async fn discover_services_with_uuid(&self, _uuid: Uuid) -> Result<Vec<Service>> {
-        self.services().await
+    pub async fn discover_services_with_uuid(&self, uuid: Uuid) -> Result<Vec<Service>> {
+        Ok(self
+            .services()
+            .await?
+            .into_iter()
+            .filter(|x| x.uuid() == uuid)
+            .collect())
     }
 
     /// Get previously discovered services.
     ///
-    /// If no services have been discovered yet, this method may either perform service discovery or return an error.
+    /// If no services have been discovered yet, this method will perform service discovery.
     pub async fn services(&self) -> Result<Vec<Service>> {
-        Ok(self.inner.services().await?.into_iter().map(Service::new).collect())
+        Ok(self
+            .inner
+            .services()
+            .await?
+            .into_iter()
+            .map(|x| Service::new(self.inner.clone(), x))
+            .collect())
     }
 
-    /// Asynchronously blocks until a GATT services changed packet is received
-    pub async fn services_changed(&self) -> Result<()> {
-        let services = self.services().await?;
-        for service in services {
-            if service.uuid_async().await? == btuuid::services::GENERIC_ATTRIBUTE {
-                for characteristic in service.characteristics().await? {
-                    if characteristic.uuid_async().await? == btuuid::characteristics::SERVICE_CHANGED {
-                        let notifications = characteristic.notify().await?;
-                        pin!(notifications);
-                        return match notifications.next().await {
-                            Some(Ok(_)) => Ok(()),
-                            Some(Err(err)) => Err(err),
-                            None => Err(Error::new(
-                                ErrorKind::Internal,
-                                None,
-                                "service changed notifications ended unexpectedly".to_string(),
-                            )),
-                        };
+    /// Monitors the device for services changed events.
+    pub async fn service_changed_indications(
+        &self,
+    ) -> Result<impl Stream<Item = Result<ServicesChanged>> + Send + Unpin + '_> {
+        let mut characteristic = Err(Error::from(ErrorKind::NotFound));
+        {
+            let services = self.inner.services().await?;
+            for service in services {
+                if service.uuid().await? == btuuid::services::GENERIC_ATTRIBUTE {
+                    for c in service.characteristics().await? {
+                        if c.uuid().await? == btuuid::characteristics::SERVICE_CHANGED {
+                            characteristic = Ok(c);
+                        }
                     }
                 }
             }
         }
 
-        Err(ErrorKind::NotFound.into())
+        let notifications = Box::pin(characteristic?.notify().await?);
+        Ok(notifications.map(|data| {
+            if data.len() == 4 {
+                let start_handle = u16::from_le_bytes(data[..2].try_into().unwrap());
+                let end_handle = u16::from_le_bytes(data[2..].try_into().unwrap());
+                Ok(ServicesChanged(ServicesChangedImpl(start_handle..=end_handle)))
+            } else {
+                Err(ErrorKind::InvalidParameter.into())
+            }
+        }))
     }
 
     /// Get the current signal strength from the device in dBm.
@@ -253,5 +290,23 @@ impl DeviceImpl {
             tx_power_level,
             is_connectable,
         }
+    }
+
+    pub async fn open_l2cap_channel(
+        &self,
+        _psm: u16,
+        _secure: bool,
+    ) -> std::prelude::v1::Result<(L2capChannelReader, L2capChannelWriter), crate::Error> {
+        Err(ErrorKind::NotSupported.into())
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
+pub struct ServicesChangedImpl(std::ops::RangeInclusive<u16>);
+
+impl ServicesChangedImpl {
+    pub fn was_invalidated(&self, service: &Service) -> bool {
+        let service_id = service.0.inner.id();
+        self.0.contains(&service_id)
     }
 }